refactored some files
