+++ app/src/main/java/muna/munaleagueandroidapp/fragments/AboutFragment.java
@@ -1,4 +1,4 @@
-package muna.munaleagueandroidapp;
+package muna.munaleagueandroidapp.fragments;
 
 
 import android.os.Bundle;
@@ -7,18 +7,19 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
+import muna.munaleagueandroidapp.R;
 
-public class HomeFragment extends Fragment {
+
+public class AboutFragment extends Fragment {
 
 
-    public HomeFragment() {
-
+    public AboutFragment() {
     }
 
 
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
-        return inflater.inflate(R.layout.fragment_home, container, false);
+        return inflater.inflate(R.layout.fragment_about, container, false);
     }
 
 }
