+++ app/src/main/java/muna/munaleagueandroidapp/networking/HttpHandler.java
@@ -1,4 +1,4 @@
-package muna.munaleagueandroidapp;
+package muna.munaleagueandroidapp.networking;
 
 import android.util.Log;
 import java.io.BufferedInputStream;
