+++ app/src/main/res/layout/fragment_home.xml
@@ -2,7 +2,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="muna.munaleagueandroidapp.HomeFragment">
+    tools:context="muna.munaleagueandroidapp.fragments.HomeFragment">
 
 
     <TextView
