+++ app/src/main/res/layout/team_challenge_list.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:background="#0e2845"
 
-    tools:context="muna.munaleagueandroidapp.TeamChallengesFragment"
+    tools:context="muna.munaleagueandroidapp.fragments.TeamChallengesFragment"
     android:orientation="horizontal">
